requirements.txt
numpy==1.21.2
pdf2image==1.16.0
Pillow==8.3.2